Function dbg_rust::working::prettier

pub fn prettier(child: Pid)

Handle process stopping events and print information when a SIGTRAP signal is received.

This function continuously waits for the child process to stop and checks if it’s due to a SIGTRAP signal, indicating a breakpoint hit. When a SIGTRAP is detected, it prints information about it and then breaks out of the loop.

Arguments

child - The process ID (Pid) of the child being debugged.

Panics

This function panics if it fails to get the register states of the child process.